now debug is sconnected from release

+++ migrations/2023-01-25-075527_marker_types/up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS public.marker_types
 (
     id BIGSERIAL NOT NULL,
     name TEXT COLLATE pg_catalog."default" NOT NULL,
-    points FLOAT NOT NULL,
+    points DOUBLE PRECISION NOT NULL,
     CONSTRAINT markers_types_pkey PRIMARY KEY (id)
 )
