+++ migrations/2023-01-25-075528_users/up.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS public.users
     email character varying(254) COLLATE pg_catalog."default" NOT NULL,
     password character varying(255) COLLATE pg_catalog."default" NOT NULL,
     is_admin boolean DEFAULT 'false',
-    points BIGINT DEFAULT 0,
+    points DOUBLE PRECISION DEFAULT 0.0,
     CONSTRAINT users_pkey PRIMARY KEY (id),
     CONSTRAINT users_email_key UNIQUE (email)
 )
